refactor: add WalletRepository and fix WalletJpaRepository and update references
