+++ src/main/java/com/payflow/application/command/auth/RegisterCommandHandler.java
@@ -9,9 +9,9 @@ import com.payflow.domain.model.user.EmailAlreadyRegisteredException;
 import com.payflow.domain.model.user.User;
 import com.payflow.domain.model.user.UserStatus;
 import com.payflow.domain.model.wallet.Wallet;
+import com.payflow.domain.repository.WalletRepository;
 import com.payflow.infrastructure.persistence.jpa.RefreshTokenRepository;
 import com.payflow.infrastructure.persistence.jpa.UserRepository;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
