+++ src/main/java/com/payflow/application/command/wallet/CreateWalletCommandHandler.java
@@ -3,7 +3,7 @@ package com.payflow.application.command.wallet;
 import com.payflow.api.dto.response.WalletResponse;
 import com.payflow.domain.model.wallet.Wallet;
 import com.payflow.domain.model.wallet.WalletAlreadyExistsException;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
+import com.payflow.domain.repository.WalletRepository;
 
 import java.util.Currency;
 import java.util.UUID;
