+++ src/main/java/com/payflow/application/command/wallet/FreezeWalletCommandHandler.java
@@ -2,7 +2,7 @@ package com.payflow.application.command.wallet;
 
 import com.payflow.domain.model.wallet.Wallet;
 import com.payflow.domain.model.wallet.WalletNotFoundException;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
+import com.payflow.domain.repository.WalletRepository;
 
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
