+++ src/main/java/com/payflow/application/query/wallet/GetWalletBalanceQueryHandler.java
@@ -3,7 +3,8 @@ package com.payflow.application.query.wallet;
 import com.payflow.api.dto.response.BalanceResponse;
 import com.payflow.domain.model.wallet.Wallet;
 import com.payflow.domain.model.wallet.WalletNotFoundException;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
+import com.payflow.domain.repository.WalletRepository;
+import com.payflow.infrastructure.persistence.jpa.WalletJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
