+++ src/test/java/com/payflow/application/command/auth/LoginIntegrationTest.java
@@ -5,7 +5,7 @@ import com.payflow.TestcontainersConfiguration;
 import com.payflow.api.dto.request.LoginRequest;
 import com.payflow.api.dto.request.RegisterRequest;
 import com.payflow.infrastructure.persistence.jpa.UserRepository;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
+import com.payflow.infrastructure.persistence.jpa.WalletJpaRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +24,7 @@ class LoginIntegrationTest {
     @Autowired
     private RestTestClient restTestClient;
     @Autowired private UserRepository userRepository;
-    @Autowired private WalletRepository walletRepository;
+    @Autowired private WalletJpaRepository walletRepository;
 
     private String userEmail;
