+++ src/test/java/com/payflow/application/command/auth/RegisterIntegrationTest.java
@@ -3,7 +3,7 @@ package com.payflow.application.command.auth;
 import com.payflow.TestcontainersConfiguration;
 import com.payflow.api.dto.request.RegisterRequest;
 import com.payflow.infrastructure.persistence.jpa.UserRepository;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
+import com.payflow.infrastructure.persistence.jpa.WalletJpaRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,7 +19,7 @@ class RegisterIntegrationTest {
 
     @Autowired private RestTestClient restTestClient;
     @Autowired private UserRepository userRepository;
-    @Autowired private WalletRepository walletRepository;
+    @Autowired private WalletJpaRepository walletRepository;
 
     @Test
     void shouldRegisterUserCreateWalletAndReturnTokens() {
