+++ src/test/java/com/payflow/application/command/wallet/CreateWalletCommandHandlerTest.java
@@ -3,7 +3,7 @@ package com.payflow.application.command.wallet;
 import com.payflow.api.dto.response.WalletResponse;
 import com.payflow.domain.model.wallet.Wallet;
 import com.payflow.domain.model.wallet.WalletAlreadyExistsException;
-import com.payflow.infrastructure.persistence.jpa.WalletRepository;
+import com.payflow.infrastructure.persistence.jpa.WalletJpaRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -23,7 +23,7 @@ import static org.mockito.Mockito.*;
 class CreateWalletCommandHandlerTest {
 
     @Mock
-    private WalletRepository walletRepository;
+    private WalletJpaRepository walletRepository;
 
     @InjectMocks
     private CreateWalletCommandHandler handler;
