+++ src/test/java/com/payflow/application/service/RefreshTokenServiceTest.java
@@ -71,7 +71,7 @@ class RefreshTokenServiceTest {
     }
 
 
-        @Test
+    @Test
     void shouldThrowWhenTokenExpired() {
         // Given
         RefreshToken expired = RefreshToken.builder()
